moved states to own file and fixed some ts errors
